requirements.txt
channels==3.0.3
Django==3.1.7
django-recaptcha==2.0.6
python-dotenv==0.16.0
